Funcionando antes del DELETE

+++ lib/infrastructure/models/sactidb/product_details_sactidb.dart
@@ -2,7 +2,7 @@
 class ProductDetails {
     final int id;
     final String nombre;
-    final int precio;
+    final double precio;
     final DateTime fechaCreacion;
     final Categoria marca;
     final Categoria categoria;
